pyproject.toml
[project]
name = "boxautolabel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "timm>=1.0.19",
    "Pillow>=9.0.0",
    "matplotlib>=3.5.0",
    "tensorboard>=2.10.0",
    "huggingface-hub>=0.34.4",
    "requests>=2.28.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
]

[dependency-groups]
dev = [
    "huggingface-hub[cli]>=0.34.4",
    "ruff>=0.12.9",
]
